Split long source files into short ones
